feat: add shared presence service
